Implemented FDHANDLER::write().
Now all file I/O methods in FDHANDLER make FDHANDLER::error()
to return true on error.

Method chaining implemented in FDHANDLER::write().

Fixed error in both read methods of FDHANDLER where in case the file
to read was in the current working directory (this is, there is not
folders in the path), the methods tried to change to a directory
called the same as the filename.